Local_Max_Identification_Prototype_1
This is the first prototype for the project. 
Run time =.0479 sec
Further to update the code with numpy arrays and check if that reduced the Run Time. 
Also make use of OOPS to make the code more dynamic.
It also contains the functions for data cleaning process as well.
The actual input data is in the form of txt files.
All the Data cleaning and converting into dictionary format done using few functions within.